+++ .gitignore
@@ -1,0 +1,4 @@
+target/
+src-gen/
+xtend-gen/
+plugin.xml_gen
