+++ pl.mrasoft.mridl.fragments/.gitignore
@@ -1,0 +1,1 @@
+!xtend-gen/
