+++ pl.mrasoft.mridl.fragments/build.properties
@@ -1,0 +1,4 @@
+source.. = src/,\
+          xtend-gen/
+bin.includes = META-INF/,\
+               .
