+++ pl.mrasoft.mridl.fragments/src/pl/mrasoft/mridl/fragments/MridlScopingFragment.xpt
