--- pl.mrasoft.mridl.tests/.gitignore
@@ -1,1 +1,0 @@
-/bin/
