--- pl.mrasoft.mridl.tests/xtend-gen/.gitignore
@@ -1,1 +1,0 @@
-/pl/
