--- pl.mrasoft.mridl.ui/.gitignore
@@ -1,2 +1,0 @@
-/bin/
-/plugin.xml_gen
