--- pl.mrasoft.mridl.ui/src-gen/.gitignore
@@ -1,1 +1,0 @@
-/pl/
