--- pl.mrasoft.mridl.ui/xtend-gen/.gitignore
@@ -1,1 +1,0 @@
-/pl/
