+++ pl.mrasoft.mridl/.classpath
@@ -5,5 +5,6 @@
 	<classpathentry excluding=".gitignore" kind="src" path="xtend-gen"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
-	<classpathentry kind="output" path="bin"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/pl.mrasoft.mridl.fragments"/>
+	<classpathentry kind="output" path="target/classes"/>
 </classpath>
