+++ pl.mrasoft.mridl/.gitignore
@@ -1,3 +1,2 @@
-/bin/
-/plugin.xml_gen
-/.antlr-generator-3.2.0-patch.jar
+/model/
+.antlr-generator-3.2.0-patch.jar
