+++ pl.mrasoft.mridl/.settings/org.eclipse.core.resources.prefs
@@ -1,3 +1,2 @@
 eclipse.preferences.version=1
-encoding//src/pl/mrasoft/mridl/scoping/MridlScopingFragment.xpt=ISO-8859-1
 encoding/<project>=UTF-8
