--- pl.mrasoft.mridl/examplesworkspace/pl.mrasoft.mridl.examples/.gitignore
@@ -1,1 +1,0 @@
-/bin/
