--- pl.mrasoft.mridl/examplesworkspace/pl.mrasoft.mridl.examples/src/pl/mrasoft/mridl/examples/d1/AllFeatures.mridl
@@ -1,49 +1,0 @@
-@doc("global doc")
-namespace "http://mrasoft.pl/mridl/examples/d1/AllFeatures"
-
-import "Imported 2.mridl" as i
-import ../d2/Imported.mridl as i2
-
-@doc("op1 documentation")
-operation
-	@doc("a doc") int[] a op1(@doc("f doc") string? f, TestComplex2 d)
-	throws @doc("fault doc1") TestFault, @doc("fault doc2") i:TestComplexType
-operation void op2()
-
-complex TestFault {
-
-
-	i2:TestComplexType2 b
-	TestEnum e
-	i:TestComplexType s
-	@doc("d docume
-	ntation")
-	i:ImportedTestSimple d
-}
-
-complex TestComplex2 extends i:TestComplexType {
-
-	i2:TestComplexType2[1..*] b
-	int[1..2] c
-	string(2) s
-	string s2
-	string(2)[] s3
-	decimal(2, 1) d
-
-}
-
-simple TestSimple2 {
-	base = string
-	pattern = "[a-zA-Z][a-zA-Z][a-zA-Z]"
-}
-
-simple TestSimple3 {
-	base = i:ImportedTestSimple
-	pattern = "[a-zA-Z][a-zA-Z][a-zA-Z]"
-}
-
-enum TestEnum {
-	@doc("ENUM1 documentation") ENUM1
-	ENUM2
-} 
-
