--- pl.mrasoft.mridl/examplesworkspace/pl.mrasoft.mridl.examples/src/pl/mrasoft/mridl/examples/d1/Imported 2.mridl
@@ -1,13 +1,0 @@
-namespace "http://mrasoft.pl/mridl/examples/d1/Imported"
-
-complex TestComplexType {
-
-	string aaa
-	int bbb
-
-}
-
-simple ImportedTestSimple {
-	base = string
-	pattern = "[a-zA-Z][a-zA-Z][a-zA-Z]"
-}
