--- pl.mrasoft.mridl/examplesworkspace/pl.mrasoft.mridl.examples/src/pl/mrasoft/mridl/examples/d2/Imported.mridl
@@ -1,8 +1,0 @@
-namespace "http://mrasoft.pl/mridl/examples/d2/Imported"
-
-complex TestComplexType2 {
-
-	string aaa
-	int bbb
-
-}
