--- pl.mrasoft.mridl/model/.gitignore
@@ -1,1 +1,0 @@
-/generated/
