--- pl.mrasoft.mridl/src-gen/.gitignore
@@ -1,1 +1,0 @@
-/pl/
