--- pl.mrasoft.mridl/xtend-gen/.gitignore
@@ -1,1 +1,0 @@
-/pl/
